Filter: Changed Cubic Filter algorith in Filter to look only at the middle +/- 10 Clock tics for the main zero crossing
LendaEvent:
added vector for the CFDSourceTrigBits
added the clear statment and the pushing method

Packer:
Packer now calculates time without the sourceTrigBit.  This seems to work.
The method being used for the softwareCFD extract seems to already take this
phase into account.  Also changed the RePackSoftwareTime method to use the same
formula.

Added the TimmingAnalyzer Program to the repo.